Completed prc 7

+++ ClassLibrary/clsStock.cs
@@ -11,7 +11,7 @@ namespace ClassLibrary
         public int GameId  { get; set; }
         public string GameTitle { get; set; }
         public DateTime GameReleaseDate { get; set; }
-        public int GamePrice { get; set; }
+        public decimal  GamePrice { get; set; }
         public int StockQty { get; set; }
         public int GameRating { get; set; }
         public bool IsDigital { get; set; }
